adding changes made by antigravity

+++ pkg/dsm/webapi/iscsi.go
@@ -200,7 +200,7 @@ func (dsm *DSM) LunCreate(spec LunCreateSpec) (string, error) {
 	params.Add("method", "create")
 	params.Add("version", "1")
 	params.Add("name", strconv.Quote(spec.Name))
-	params.Add("size", strconv.FormatInt(int64(spec.Size), 10))
+	params.Add("size", strconv.FormatInt(spec.Size, 10))
 	params.Add("type", spec.Type)
 	params.Add("location", spec.Location)
 	params.Add("description", spec.Description)
